requirement.txt
streamlit
openai
PyGithub